Decompose and organize `import-tests` script (#1165)
* Separate git operations

* separate out test parser

* Single function for test parsing

* Break test parser up

* Reapply test plan operation changes

* Use strategy pattern for test parsing, add some jsdocs

* Additional comments and use JSDoc types more

* Only export as necessary, keep V1 naming, add tmp to gitignore

* Update deprecation date comparison
